Changing projected saving to a floating point number

+++ python_introduction/finance_calculator.py
@@ -5,7 +5,7 @@ monthly_savings = int(income) - int(expenses)
 yearly_savings = monthly_savings * 12
 rate = 0.05
 
-projected_savings = yearly_savings + (yearly_savings * rate)
+projected_savings = float(yearly_savings + (yearly_savings * rate))
 
 print("Your monthly savings are $", monthly_savings)
 print("Projected savings after one year, with interest, is: $", projected_savings)
